Use nupic.bindings 0.6.3
This requires updating other dependencies. A couple tests fail with
the new numpy version, so I fixed those tests.

+++ tests/integration/nupic/algorithms/knn_classifier_test/classifier_test.py
@@ -74,7 +74,7 @@ class KNNClassifierTest(unittest.TestCase):
       patternDict[i]['pattern'] = patterns[i]
       patternDict[i]['category'] = numpy.random.randint(0, numClasses-1)
       testDict[i] = copy.deepcopy(patternDict[i])
-      testDict[i]['pattern'][:0.02*patternSize] = numpy.random.rand()
+      testDict[i]['pattern'][:int(0.02*patternSize)] = numpy.random.rand()
       testDict[i]['category'] = None
 
     LOGGER.info("\nTesting KNN Classifier with L2 norm")
